Add POM
Basic POM to be used as a parent. Includes the versions for all maven plugins
used. Setup for java 17, which is also enforced.

For now it only contain two dependencies; junit-jupiter and assertj-core.

The intention of this parent is to provide default values for other projects.
They should for the most part only have to update their parent and all the
other versions should just follow. This parent does not automatically include
any dependencies other than plugins, it's the downstream projects'
responsibility to import dependencies that it need.